pyproject.toml
[tool.poetry]
name = "maintenance-bot"
version = "0.1.0"
description = "Telegram bot for maintenance requests"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
python-telegram-bot = {extras = ["ext"], version = "^22.2"}
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
gspread = "^6.2.1"
google-api-python-client = "^2.176.0"
tenacity = "^9.1.2"
pytz = "^2025.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-mock = "^3.14.1"
ruff = "^0.12.4"
black = "^25.1.0"
pytest-asyncio = "^1.1.0"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = ["E", "W", "F", "I", "C", "B"]
ignore = ["E501"]

[tool.black]
line-length = 88
